+++ DependencyInjection/EkinoNewRelicExtension.php
@@ -11,7 +11,6 @@
 
 namespace  Ekino\Bundle\NewRelicBundle\DependencyInjection;
 
-use Monolog\Handler\NewRelicHandler;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
@@ -36,14 +35,8 @@ class EkinoNewRelicExtension extends Extension
         $loader = new Loader\XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.xml');
 
-        $container->getDefinition('ekino.new_relic.request_listener')
-            ->replaceArgument(2, $config['ignored_routes'])
-            ->replaceArgument(3, $config['ignored_paths']);
-
-        $container->getDefinition('ekino.new_relic.command_listener')
-            ->replaceArgument(2, $config['ignored_commands']);
-
         if (!$config['enabled']) {
+            $config['monolog']['enabled'] = false;
             $interactor = 'ekino.new_relic.interactor.blackhole';
         } elseif (isset($config['interactor'])) {
             $interactor = $config['interactor'];
@@ -62,22 +55,7 @@ class EkinoNewRelicExtension extends Extension
             $container->setAlias('ekino.new_relic.interactor', $interactor);
         }
 
-        $container->getDefinition('ekino.new_relic.response_listener')
-            ->replaceArgument(2, $config['instrument'])
-            ->replaceArgument(3, $config['using_symfony_cache'])
-        ;
-
-        if (!$config['log_exceptions']) {
-            $container->removeDefinition('ekino.new_relic.exception_listener');
-        }
-
-        $container->setParameter('ekino.new_relic.log_deprecations', $config['log_deprecations']);
-
-        if (!$config['log_commands']) {
-            $container->removeDefinition('ekino.new_relic.command_listener');
-        }
-
-        if (!$config['deployment_names']) {
+        if (!empty($config['deployment_names'])) {
             $config['deployment_names'] = array_values(array_filter(explode(';', $config['application_name'])));
         }
 
@@ -89,49 +67,79 @@ class EkinoNewRelicExtension extends Extension
             ->replaceArgument(4, $config['deployment_names'])
         ;
 
-        if (!$config['enabled']) {
-            $config['log_logs']['enabled'] = false;
+        if ($config['http']['enabled']) {
+            $loader->load('http_listener.xml');
+            $container->getDefinition('ekino.new_relic.request_listener')
+                ->replaceArgument(2, $config['http']['ignored_routes'])
+                ->replaceArgument(3, $config['http']['ignored_paths'])
+                ->replaceArgument(4, $this->getTransactionNamingService($config))
+                ->replaceArgument(5, $config['http']['using_symfony_cache']);
+
+            $container->getDefinition('ekino.new_relic.response_listener')
+                ->replaceArgument(2, $config['http']['instrument'])
+                ->replaceArgument(3, $config['http']['using_symfony_cache']);
         }
-        $container->setParameter('ekino.new_relic.log_logs', $config['log_logs']);
-        if ($config['log_logs']['enabled']) {
-            if (!class_exists(NewRelicHandler::class)) {
-                throw new \LogicException('The "symfony/monolog-bundle" package must be installed in order to use "log_logs" option.');
+
+        if ($config['commands']['enabled']) {
+            $loader->load('command_listener.xml');
+            $container->getDefinition('ekino.new_relic.command_listener')
+                ->replaceArgument(2, $config['commands']['ignored_commands']);
+        }
+
+        if ($config['exceptions']['enabled']) {
+            $loader->load('exception_listener.xml');
+        }
+
+        if ($config['deprecations']['enabled']) {
+            $loader->load('deprecation_listener.xml');
+        }
+
+        if ($config['twig']) {
+            $loader->load('twig.xml');
+        }
+
+        if ($config['monolog']['enabled']) {
+            if (!class_exists(\Monolog\Handler\NewRelicHandler::class)) {
+                throw new \LogicException('The "symfony/monolog-bundle" package must be installed in order to use "monolog" option.');
             }
+            $loader->load('monolog.xml');
+            $container->setParameter('ekino.new_relic.monolog.channels', $config['monolog']['channels']);
+            $container->setAlias('ekino.new_relic.logs_handler', $config['monolog']['service']);
 
-            $container->setAlias('ekino.new_relic.logs_handler', $config['log_logs']['service']);
-
-            $level = $config['log_logs']['level'];
+            $level = $config['monolog']['level'];
             $container->findDefinition('ekino.new_relic.logs_handler')
                 ->replaceArgument(0, is_int($level) ? $level : constant('Monolog\Logger::'.strtoupper($level)))
                 ->replaceArgument(2, $config['application_name']);
         }
+    }
 
-        switch ($config['transaction_naming']) {
+    private function getTransactionNamingService(array $config): string
+    {
+        switch ($config['http']['transaction_naming']) {
             case 'controller':
-                $transaction_naming_service = new Reference('ekino.new_relic.transaction_naming_strategy.controller');
+                $serviceId = new Reference('ekino.new_relic.transaction_naming_strategy.controller');
                 break;
             case 'route':
-                $transaction_naming_service = new Reference('ekino.new_relic.transaction_naming_strategy.route');
+                $serviceId = new Reference('ekino.new_relic.transaction_naming_strategy.route');
                 break;
             case 'service':
-                if (!isset($config['transaction_naming_service']))
-                {
-                    throw new \LogicException('When using the "service", transaction naming scheme, the "transaction_naming_service" config parameter must be set.');
+                if (!isset($config['http']['transaction_naming_service'])) {
+                    throw new \LogicException(
+                        'When using the "service", transaction naming scheme, the "transaction_naming_service" config parameter must be set.'
+                    );
                 }
 
-                $transaction_naming_service = new Reference($config['transaction_naming_service']);
+                $serviceId = new Reference($config['http']['transaction_naming_service']);
                 break;
             default:
-                throw new \InvalidArgumentException(sprintf('Invalid transaction naming scheme "%s", must be "route", "controller" or "service".', $config['transaction_naming']));
+                throw new \InvalidArgumentException(
+                    sprintf(
+                        'Invalid transaction naming scheme "%s", must be "route", "controller" or "service".',
+                        $config['http']['transaction_naming']
+                    )
+                );
         }
 
-        $container->getDefinition('ekino.new_relic.request_listener')
-            ->replaceArgument(4, $transaction_naming_service)
-            ->replaceArgument(5, $config['using_symfony_cache'])
-        ;
-
-        if (!$config['twig']) {
-            $container->removeDefinition('ekino.new_relic.twig.new_relic_extension');
-        }
+        return $serviceId;
     }
 }