+++ Tests/DependencyInjection/ConfigurationTest.php
@@ -9,7 +9,7 @@ use Symfony\Component\Config\Definition\PrototypedArrayNode;
 
 class ConfigurationTest extends TestCase
 {
-    public function testIgnoredRoutes ()
+    public function testIgnoredRoutes()
     {
         $configuration = new Configuration();
         $rootNode = $configuration->getConfigTreeBuilder()
@@ -17,7 +17,7 @@ class ConfigurationTest extends TestCase
         $children = $rootNode->getChildren();
 
         /** @var PrototypedArrayNode $ignoredRoutesNode */
-        $ignoredRoutesNode = $children['ignored_routes'];
+        $ignoredRoutesNode = $children['http']->getChildren()['ignored_routes'];
 
 
         $this->assertInstanceOf('\Symfony\Component\Config\Definition\PrototypedArrayNode', $ignoredRoutesNode);
@@ -29,7 +29,7 @@ class ConfigurationTest extends TestCase
         $this->assertEquals(array('ignored_route1', 'ignored_route2'), $ignoredRoutesNode->merge(array('ignored_route1'), array('ignored_route2')));
     }
 
-    public function testIgnoredPaths ()
+    public function testIgnoredPaths()
     {
         $configuration = new Configuration();
         $rootNode = $configuration->getConfigTreeBuilder()
@@ -37,7 +37,7 @@ class ConfigurationTest extends TestCase
         $children = $rootNode->getChildren();
 
         /** @var PrototypedArrayNode $ignoredPathsNode */
-        $ignoredPathsNode = $children['ignored_paths'];
+        $ignoredPathsNode = $children['http']->getChildren()['ignored_paths'];
 
 
         $this->assertInstanceOf('\Symfony\Component\Config\Definition\PrototypedArrayNode', $ignoredPathsNode);
@@ -57,7 +57,7 @@ class ConfigurationTest extends TestCase
         $children = $rootNode->getChildren();
 
         /** @var PrototypedArrayNode $ignoredCommandsNode */
-        $ignoredCommandsNode = $children['ignored_commands'];
+        $ignoredCommandsNode = $children['commands']->getChildren()['ignored_commands'];
 
 
         $this->assertInstanceOf('\Symfony\Component\Config\Definition\PrototypedArrayNode', $ignoredCommandsNode);
@@ -75,12 +75,12 @@ class ConfigurationTest extends TestCase
 
         $config = $processor->processConfiguration(new Configuration(), array());
 
-        $this->assertEmpty($config['ignored_routes']);
-        $this->assertInternalType('array', $config['ignored_routes']);
-        $this->assertEmpty($config['ignored_paths']);
-        $this->assertInternalType('array', $config['ignored_paths']);
-        $this->assertEmpty($config['ignored_commands']);
-        $this->assertInternalType('array', $config['ignored_commands']);
+        $this->assertEmpty($config['http']['ignored_routes']);
+        $this->assertInternalType('array', $config['http']['ignored_routes']);
+        $this->assertEmpty($config['http']['ignored_paths']);
+        $this->assertInternalType('array', $config['http']['ignored_paths']);
+        $this->assertEmpty($config['commands']['ignored_commands']);
+        $this->assertInternalType('array', $config['commands']['ignored_commands']);
         $this->assertEmpty($config['deployment_names']);
         $this->assertInternalType('array', $config['deployment_names']);
     }
@@ -142,9 +142,9 @@ class ConfigurationTest extends TestCase
     {
         $processor = new Processor();
 
-        $config = $processor->processConfiguration(new Configuration(), array('ekino_new_relic' => array('ignored_routes' => $ignoredRoutesConfig)));
+        $config = $processor->processConfiguration(new Configuration(), array('ekino_new_relic' => array('http'=>array('ignored_routes' => $ignoredRoutesConfig))));
 
-        $this->assertEquals($expected, $config['ignored_routes']);
+        $this->assertEquals($expected, $config['http']['ignored_routes']);
     }
 
     /**
@@ -154,9 +154,9 @@ class ConfigurationTest extends TestCase
     {
         $processor = new Processor();
 
-        $config = $processor->processConfiguration(new Configuration(), array('ekino_new_relic' => array('ignored_paths' => $ignoredPathsConfig)));
+        $config = $processor->processConfiguration(new Configuration(), array('ekino_new_relic' => array('http'=>array('ignored_paths' => $ignoredPathsConfig))));
 
-        $this->assertEquals($expected, $config['ignored_paths']);
+        $this->assertEquals($expected, $config['http']['ignored_paths']);
     }
 
     /**
@@ -166,8 +166,8 @@ class ConfigurationTest extends TestCase
     {
         $processor = new Processor();
 
-        $config = $processor->processConfiguration(new Configuration(), array('ekino_new_relic' => array('ignored_commands' => $ignoredCommandsConfig)));
+        $config = $processor->processConfiguration(new Configuration(), array('ekino_new_relic' => array('commands'=>array('ignored_commands' => $ignoredCommandsConfig))));
 
-        $this->assertEquals($expected, $config['ignored_commands']);
+        $this->assertEquals($expected, $config['commands']['ignored_commands']);
     }
 }