+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## v2.0.0
 
-### Addded
+### Added
 
 - All functions provided by the NewRelic PHP extension are now supported in the `NewRelicInteractorInterface`.
-- Added a new `log_deprecations` parameter to logs `E_USER_DEPRECATED`. 
+- Added a new `deprecations` parameter to logs `E_USER_DEPRECATED`. 
 
 ### Changed